Introduce "filter solvers"
We had solvers for the different kind of resources but, as we only had
one type of filter, not for the filters.
We'll soon include the management of `SliceFilter` in the solvers, and
the retrieval is done in a totally different way than a default filter.

So this commit had concept of filter solvers, with only one for now,
`FilterSolver`.

Note that we added a `solve_filter` method to the registry, so for
consistency, the `solve` method of the registry was renamed
`solve_resource`.