List resources accepts more kind of resources
It can now accepts one or many resources and behave like this:

- if only one resource, return a list with the result of this resource
for each iteration of the value
- if more that one resource, return a list with, for each iteration, a
list of each subresource for the value

This means that this breaks compatibility:

Previously, `[foo, bar]` returned a list of dict with two keys, `foo`,
and `bar`.

Not it returns a list of list, each sublist having the value for `foo`
and `bar`.

To have the same behavior as before, you should write `[{foo, bar}]`.

This allows to reuse the concept of fields and objects and adds the
possibility to return list of things, not only list of dicts.