Make functions work in EntryPoints
Before this commit, as a class is created on the fly when calling
``EntryPoints``, all functions where saved as method of this class, so
they all expected ``self`` as the first argument.

This commit transform each function in staticmethod, so they can be
called on the class without having to pass ``self`` or ``cls``.